Refactor FilterEditor::Load to use static_cast for filter bounds
Updated the Load method in FilterEditor to explicitly cast filter bounds to qint64, ensuring type safety and clarity in the code. This change enhances the robustness of the filter loading process.

+++ app/src/main_window.cpp
@@ -1258,7 +1258,11 @@ void MainWindow::AddFilter(
                 delete filterEditor;
                 return;
             }
-            filterEditor->Load(resolvedFilter->row, *resolvedFilter->filterBegin, *resolvedFilter->filterEnd);
+            filterEditor->Load(
+                resolvedFilter->row,
+                static_cast<qint64>(*resolvedFilter->filterBegin),
+                static_cast<qint64>(*resolvedFilter->filterEnd)
+            );
         }
         else if (resolvedFilter->type == loglib::LogConfiguration::LogFilter::Type::Enumeration)
         {
